# public/openapi.yaml
openapi: 3.0.0

info:
  title: Jua Kali Connect API
  version: 1.0.0

servers:
  - url: http://localhost:5000/api

paths:
  /bookings:
    post:
      summary: Create a new booking for a selected artisan and service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - artisan_name
                - service
              properties:
                artisan_name:
                  type: string
                  example: "Sean"
                service:
                  type: string
                  example: "Plumbing"
                status:
                  type: string
                  enum:
                    - pending
                    - confirmed
                    - completed
                  example: "pending"

      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 4
                  artisan_name:
                    type: string
                    example: "Sean"
                  service:
                    type: string
                    example: "Plumbing"
                  status:
                    type: string
                    example: "pending"

        "400":
          description: Invalid request — missing or invalid fields

  /sessions:
    post:
      summary: Create an authenticated session for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "seanochieng@gmail.com"
                password:
                  type: string
                  format: password
                  example: "correcthorsebatterystaple"

      responses:
        "201":
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  userId:
                    type: integer
                    example: 12

        "400":
          description: Invalid request — missing or invalid fields

        "401":
          description: Invalid email or password

  /artisans:
    get:
      summary: Return artisans matching the search term and optional location filter
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            example: "seanochieng"

        - name: location
          in: query
          required: false
          schema:
            type: string
            example: "Nairobi"

      responses:
        "200":
          description: A list of matching artisans
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 3
                    name:
                      type: string
                      example: "Sean"
                    service:
                      type: string
                      example: "Plumbing"
                    rating:
                      type: number
                      example: 4.8

        "404":
          description: No matching artisans found

  /artisans/{id}:
    get:
      summary: Return the details of a specific artisan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 3

      responses:
        "200":
          description: Artisan details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 3
                  name:
                    type: string
                    example: "Sean"
                  service:
                    type: string
                    example: "Plumbing"
                  phone:
                    type: string
                    example: "+254712345678"
                  rating:
                    type: number
                    example: 4.8

        "404":
          description: Artisan not found

  /bookings/{id}:
    patch:
      summary: Update an existing booking's information or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - pending
                    - confirmed
                    - completed
                  example: "completed"

      responses:
        "200":
          description: Booking updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Booking updated"

        "400":
          description: Invalid request

        "404":
          description: Booking not found